Cleaned up stuff theat created errors, like double db context and calling the wrong context i several files.
Also tried to make sure that the two new renting pages were visable when pulled. One of them didn't work after my git pull. Changed the names for these pages so they are easier to identify/tell apart.

+++ Models/FilmActor.cs
@@ -5,9 +5,9 @@ namespace Retro_grupp_g.Models;
 
 public partial class FilmActor
 {
-    public ushort ActorId { get; set; }
+    public int ActorId { get; set; }
 
-    public ushort FilmId { get; set; }
+    public int FilmId { get; set; }
 
     public DateTime LastUpdate { get; set; }
